Collects from database, but not unique stuff
So here’s a simple script that goes to spollywood.com 50 times, fetches
the title of the movie and the one-liner spoiler, inserts it into a new
or old table depending on if it exists or not.

Doesn’t check if it already used the movie or not, that should be next
commit. Also sometimes the spoilers or titles have messed up text,
maybe from some escape character somewhere? Idk